+++ common/src/main/java/io/github/mortuusars/exposure/ExposureClient.java
@@ -12,6 +12,7 @@ import io.github.mortuusars.exposure.client.render.photograph.PhotographStyle;
 import io.github.mortuusars.exposure.client.render.photograph.PhotographRenderer;
 import io.github.mortuusars.exposure.client.render.photograph.PhotographStyles;
 import io.github.mortuusars.exposure.client.util.Minecrft;
+import io.github.mortuusars.exposure.integration.ModCompatibilityClient;
 import io.github.mortuusars.exposure.world.camera.capture.CaptureType;
 import io.github.mortuusars.exposure.world.item.camera.Attachment;
 import io.github.mortuusars.exposure.world.photograph.PhotographType;
@@ -57,6 +58,8 @@ public class ExposureClient {
         cycles().addParallelTask(new ClearStaleRenderedImagesIndefiniteTask());
 
         registerItemModelProperties();
+
+        ModCompatibilityClient.init();
     }
 
     public static Cycles cycles() {