+++ common/src/main/java/io/github/mortuusars/exposure/client/capture/action/ForceRegularOrSelfieCameraTypeAction.java
@@ -8,8 +8,8 @@ import net.minecraft.client.Minecraft;
 import org.jetbrains.annotations.Nullable;
 
 public class ForceRegularOrSelfieCameraTypeAction implements CaptureAction {
-    private final CameraHolder holder;
-    private CameraType cameraTypeBeforeCapture = CameraType.FIRST_PERSON;
+    protected final CameraHolder holder;
+    protected CameraType cameraTypeBeforeCapture = CameraType.FIRST_PERSON;
 
     public ForceRegularOrSelfieCameraTypeAction(CameraHolder holder) {
         this.holder = holder;