+++ common/src/main/java/io/github/mortuusars/exposure/integration/Mods.java
@@ -4,6 +4,8 @@ import io.github.mortuusars.exposure.PlatformHelper;
 
 public class Mods {
     public static final Mod CREATE = new Mod("create");
+    public static final Mod REAL_CAMERA = new Mod("realcamera");
+    public static final Mod SHOULDER_SURFING = new Mod("shouldersurfing");
 
     public record Mod(String id) {
         public boolean isLoaded() {