+++ neoforge/src/main/java/io/github/mortuusars/exposure/neoforge/ExposureNeoForgeClient.java
@@ -1,6 +1,5 @@
 package io.github.mortuusars.exposure.neoforge;
 
-import io.github.mortuusars.exposure.integration.ModCompatibilityClient;
 import net.neoforged.fml.ModContainer;
 import net.neoforged.neoforge.client.gui.ConfigurationScreen;
 import net.neoforged.neoforge.client.gui.IConfigScreenFactory;
@@ -8,6 +7,5 @@ import net.neoforged.neoforge.client.gui.IConfigScreenFactory;
 public class ExposureNeoForgeClient {
     public static void init(ModContainer modContainer) {
         modContainer.registerExtensionPoint(IConfigScreenFactory.class, ConfigurationScreen::new);
-        ModCompatibilityClient.handle();
     }
 }